% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath_contains}
\alias{xpath_contains}
\title{Creates an XPath expression that checks if a node contains any of the
specified tags.}
\usage{
xpath_contains(tags)
}
\arguments{
\item{tags}{A character vector of tags to check for.}
}
\value{
A string representing the XPath expression.
}
\description{
Creates an XPath expression that checks if a node contains any of the
specified tags.
}
\examples{
xpath_contains(c("class", "id"))

}
