% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath_exclude_attrs}
\alias{xpath_exclude_attrs}
\title{Exclude XPath attributes from a node-set}
\usage{
xpath_exclude_attrs(attrs)
}
\arguments{
\item{attrs}{character vector of XPath attributes to exclude}
}
\value{
XPath expression to filter out nodes with specified attributes
}
\description{
Filters out nodes that have certain attributes from a node-set.
}
\examples{
xpath_exclude_attrs(c("class", "id"))

}
