% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath_get_index}
\alias{xpath_get_index}
\title{Get index for a certain XPath node}
\usage{
xpath_get_index(root, nodeset)
}
\arguments{
\item{root}{A string containing a full XPath that serves as the root for the
\code{nodeset}.}

\item{nodeset}{A string vector containing a XPath nodeset. The index will be
calculated based on the order of the nodes in this nodeset.}
}
\value{
A string representing the XPath expression for the index.
}
\description{
Creates XPath expression that returns the index of a specific node within
a node-set.
}
\examples{
root <- "//table/tr"
nodeset <- "//table/tr/td"
xpath_get_index(root, nodeset)

}
