% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath_include_attrs}
\alias{xpath_include_attrs}
\title{Include XPath attributes from a node-set}
\usage{
xpath_include_attrs(attrs)
}
\arguments{
\item{attrs}{a character vector of XPath attributes to include}
}
\value{
a string representing the resulting XPath expression#'
}
\description{
Creates an XPath expression to filter nodes that have the specified
attributes.
}
\examples{
# Filter out `class` and `id` attributes from a node-set
xpath_include_attrs(c("class", "id"))

}
