% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath_intersect_nodesets}
\alias{xpath_intersect_nodesets}
\title{XPath node-set intersection}
\usage{
xpath_intersect_nodesets(start, end = NULL)
}
\arguments{
\item{start}{a string containing the XPath element used as starting point}

\item{end}{a string containing the XPath element used as end point.}
}
\value{
a string representing the resulting XPath expression
}
\description{
Finds XPath nodes that are between two node-sets of the same kind using the
Kayessian method for node-set intersection.
}
\details{
The Kayessian XPath expression is:

\verb{$ns1[count(.|$ns2) = count($ns2)]}

Where \verb{$ns1} and \verb{$ns2} are the node-set 1 and node-set 2 respectively.

If \code{end} is NULL, the next sibling node of the same type of the starting
node-set will be used.
}
\examples{
start <- "//table/tr"
end <- "//table/tr/td"
xpath_intersect_nodesets(start, end)

}
